requirements.txt
torch
torchvision
opencv2
